SEE TODO: Use HashTreePMap in PlanNodeStatsEstimate to reduce map copying
Previously there was a lot of map copying (through ImmutableMap.copyOf(...)
and new HashMap(...)) which was significantly impacting stats code performance.
HashTreePMap is much better for cases where individual entries of base map
are modified which is common case in stats code.

TODO: This should be split into fixups. Keeping as commit for now since
it is one integral change