Propagate unknown expression stats filtering from leaf
If FilterStatsCalculator encounters unknown (unsupported) expression
that information is passed up the tree. This allows us to not reverse
infer information from stats obtained heuristically for unsupported
expression. Which was biggest problem for NOT above unsupported
expression.

The only expression which may not propagate unsupported expression flag
higher the tree is logical AND. Which in case one of the branches could
be properly computed, uses that statistics. As the other branch is based
on unsupported expression that value is yet filtered out by heuristic
factor.